requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
redis==4.6.0
scikit-learn==1.2.2
pydantic==2.3.0
numpy==1.24.3
pandas==2.1.0
python-dotenv==1.0.0
prometheus-client==0.17.1
loguru==0.7.0